Change Donut3DChart into DonutChart
Drop the old 2d donut chart and change the 3d donut chart into the new
donut chart.
